+++ test/lib/glimmer/util/url_builder_test.rb
@@ -1,0 +1,199 @@
+require './test/test_helper'
+
+require 'glimmer/util/url_builder'
+
+describe Glimmer::Util::UrlBuilder do
+  subject do
+    Glimmer::Util::UrlBuilder.new
+  end
+  
+  it 'builds URL with specified host' do
+    url = subject.host('www.google.com').to_url
+    _(url).must_equal 'https://www.google.com'
+  end
+  
+  it 'builds URL with nil host' do
+    url = subject.host(nil).to_url
+    _(url).must_equal 'https://'
+  end
+  
+  it 'builds URL with unspecified host' do
+    url = subject.to_url
+    _(url).must_equal 'https://'
+  end
+  
+  it 'builds URL with specified scheme and host' do
+    url = subject.scheme('http').host('www.google.com').to_s
+    _(url).must_equal 'http://www.google.com'
+  end
+  
+  it 'builds URL with specified scheme and host where schema includes ://' do
+    url = subject.scheme('http://').host('www.google.com').build
+    _(url).must_equal 'http://www.google.com'
+  end
+  
+  it 'builds URL with specified scheme, host, and path' do
+    url = subject.scheme('http').host('www.google.com').path('/results').to_url
+    _(url).must_equal 'http://www.google.com/results'
+  end
+  
+  it 'builds URL with specified scheme, host, and path ending with .json' do
+    url = subject.scheme('http').host('www.google.com').path('/results.json').to_url
+    _(url).must_equal 'http://www.google.com/results.json'
+  end
+  
+  it 'builds URL with specified scheme, host, and nil path' do
+    url = subject.scheme('http').host('www.google.com').path(nil).to_url
+    _(url).must_equal 'http://www.google.com'
+  end
+  
+  it 'builds URL with specified scheme, host, and empty path (meaning /)' do
+    url = subject.scheme('http').host('www.google.com').path('').to_url
+    _(url).must_equal 'http://www.google.com/'
+  end
+  
+  it 'builds URL with specified scheme, host, and path containing empty spaces only' do
+    url = subject.scheme('http').host('www.google.com').path('  ').to_url
+    _(url).must_equal 'http://www.google.com/'
+  end
+  
+  it 'builds URL with specified scheme, host, and path without forward slash at the beginning' do
+    url = subject.scheme('http').host('www.google.com').path('results').to_url
+    _(url).must_equal 'http://www.google.com/results'
+  end
+  
+  it 'builds URL with specified scheme, host, path, and param' do
+    url = subject.scheme('http').host('www.google.com').path('/results').param('q', 'sports').to_url
+    _(url).must_equal 'http://www.google.com/results?q=sports'
+  end
+  
+  it 'builds URL with specified scheme, host, path, and nil param' do
+    url = subject.scheme('http').host('www.google.com').path('/results').param('q', nil).to_url
+    _(url).must_equal 'http://www.google.com/results'
+  end
+  
+  it 'builds URL with specified scheme, host, path, and empty param' do
+    url = subject.scheme('http').host('www.google.com').path('/results').param('q', '').to_url
+    _(url).must_equal 'http://www.google.com/results'
+  end
+  
+  it 'builds URL with specified scheme, host, path, and param that becomes empty on a later param call' do
+    url = subject.scheme('http').host('www.google.com').path('/results').param('q', 'sports').param('q', '').to_url
+    _(url).must_equal 'http://www.google.com/results'
+  end
+  
+  it 'builds URL with specified scheme, host, path, and 2 params specified one by one' do
+    url = subject.scheme('http').host('www.google.com').path('/results').param('q', 'sports').param('year', '2039').to_url
+    _(url).must_equal 'http://www.google.com/results?q=sports&year=2039'
+  end
+  
+  it 'builds URL with specified scheme, host, path, and 2 params specified by a hash of string keys' do
+    url = subject.scheme('http').host('www.google.com').path('/results').params('q' => 'sports', 'year' => '2039').to_url
+    _(url).must_equal 'http://www.google.com/results?q=sports&year=2039'
+  end
+  
+  it 'builds URL with specified scheme, host, path, and 2 params specified by a hash of symbol keys' do
+    url = subject.scheme('http').host('www.google.com').path('/results').params(q: 'sports', year: '2039').to_url
+    _(url).must_equal 'http://www.google.com/results?q=sports&year=2039'
+  end
+  
+  it 'builds URL with specified scheme, host, path, and query' do
+    url = subject.scheme('http').host('www.google.com').path('/results').query('q=sports&year=2039').to_url
+    _(url).must_equal 'http://www.google.com/results?q=sports&year=2039'
+  end
+  
+  it 'builds URL with specified scheme, host, path, query, and extra param' do
+    url = subject.scheme('http').host('www.google.com').path('/results').query('q=sports&year=2039').param('month', 'apr').to_url
+    _(url).must_equal 'http://www.google.com/results?q=sports&year=2039&month=apr'
+  end
+  
+  it 'builds URL with specified scheme, host, path, query, and extra param that overrides an existing param from query' do
+    url = subject.scheme('http').host('www.google.com').path('/results').query('q=sports&year=2039').param('q', 'news').to_url
+    _(url).must_equal 'http://www.google.com/results?q=news&year=2039'
+  end
+  
+  it 'builds URL with specified scheme, host, path, param, and fragment' do
+    url = subject.scheme('http').host('www.google.com').path('/results').param('q', 'sports').fragment('#yesterday').to_url
+    _(url).must_equal 'http://www.google.com/results?q=sports#yesterday'
+  end
+  
+  it 'builds URL with specified scheme, host, path, param, and nil fragment' do
+    url = subject.scheme('http').host('www.google.com').path('/results').param('q', 'sports').fragment(nil).to_url
+    _(url).must_equal 'http://www.google.com/results?q=sports'
+  end
+  
+  it 'builds URL with specified scheme, host, path, param, and empty fragment' do
+    url = subject.scheme('http').host('www.google.com').path('/results').param('q', 'sports').fragment('').to_url
+    _(url).must_equal 'http://www.google.com/results?q=sports#'
+  end
+  
+  it 'builds URL with specified scheme, host, path, param, and fragment containing empty spaces only' do
+    url = subject.scheme('http').host('www.google.com').path('/results').param('q', 'sports').fragment('  ').to_url
+    _(url).must_equal 'http://www.google.com/results?q=sports#'
+  end
+  
+  it 'builds URL with specified scheme, host, path, param, and empty fragment starting with #' do
+    url = subject.scheme('http').host('www.google.com').path('/results').param('q', 'sports').fragment('#').to_url
+    _(url).must_equal 'http://www.google.com/results?q=sports#'
+  end
+  
+  it 'builds URL with specified scheme, host, port, path, param, and fragment' do
+    url = subject.scheme('http').host('localhost').port('3000').path('/results').param('q', 'sports').fragment('#yesterday').to_url
+    _(url).must_equal 'http://localhost:3000/results?q=sports#yesterday'
+  end
+  
+  it 'builds URL with specified scheme, host, path, param, and fragment without # at the beginning' do
+    url = subject.scheme('http').host('www.google.com').path('/results').param('q', 'sports').fragment('yesterday').to_url
+    _(url).must_equal 'http://www.google.com/results?q=sports#yesterday'
+  end
+  
+  it 'builds URL with specified URL' do
+    url = subject.url('http://www.google.com/results?q=sports#yesterday').to_url
+    _(url).must_equal 'http://www.google.com/results?q=sports#yesterday'
+  end
+  
+  it 'builds URL with specified URL with path ending in .json' do
+    url = subject.url('https://www.google.com/results.json?q=sports#yesterday').to_url
+    _(url).must_equal 'https://www.google.com/results.json?q=sports#yesterday'
+  end
+  
+  it 'builds URL with specified URL with port' do
+    url = subject.url('http://localhost:3000/results?q=sports#yesterday').to_url
+    _(url).must_equal 'http://localhost:3000/results?q=sports#yesterday'
+  end
+  
+  it 'builds URL with specified URL not containing query params' do
+    url = subject.url('http://www.google.com/results#yesterday').to_url
+    _(url).must_equal 'http://www.google.com/results#yesterday'
+  end
+  
+  it 'builds URL with specified URL not containing query params or fragment' do
+    url = subject.url('http://www.google.com/results').to_url
+    _(url).must_equal 'http://www.google.com/results'
+  end
+  
+  it 'builds URL with specified URL not containing path, query params or fragment' do
+    url = subject.url('http://www.google.com').to_url
+    _(url).must_equal 'http://www.google.com'
+  end
+  
+  it 'builds URL with specified URL not containing scheme, path, query params or fragment' do
+    url = subject.url('www.google.com').to_url
+    _(url).must_equal 'https://www.google.com'
+  end
+  
+  it 'builds URL with specified URL not containing host, path, query params or fragment' do
+    url = subject.url('http://').to_url
+    _(url).must_equal 'http://'
+  end
+  
+  it 'builds URL with specified URL not containing scheme, host, path, query params or fragment' do
+    url = subject.url('').to_url
+    _(url).must_equal 'https://'
+  end
+  
+  it 'builds URL with specified URL and an extra query parameter' do
+    url = subject.url('http://www.google.com/results?q=sports#yesterday').param('year', '2040').to_url
+    _(url).must_equal 'http://www.google.com/results?q=sports&year=2040#yesterday'
+  end
+end