+++ test/test_helper.rb
@@ -1,0 +1,4 @@
+require 'minitest'
+require 'minitest/autorun'
+require 'minitest/focus'
+require 'puts_debuggerer'